pyproject.toml
[project]
name = "the-youtuber-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.24.0",
    "fastapi>=0.122.0",
    "google-genai>=1.51.0",
    "google-generativeai>=0.8.5",
    "lancedb>=0.25.3",
    "pydantic-ai>=1.24.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
